Unsquashfs: Add -pseudo-file option
Add an option which outputs a pseudo-file representation
of the input Squashfs filesystem.

This pseudo-file can be optionally edited, and used as input
to Mksquashfs to recreate the filesystem (with perhaps different
compression type and options, or files added or excluded).

Such a pseudo file can be preferable to completely uncompressing
and unpacking into a directory hierarchy, and then re-running
Mksquashfs on that directory structure.

For directories, symlinks, devices etc (or everything not a regular
file), the pseudo definition contains all the necessary
information to recreate it.  Obviously, for regular files, this
is not the case, and the data to recreate the file needs to
come from somewhere.

Two techniques are supported by the pseudo-file option

1. Create an "F" pseudo-definition, which produces the
   regular file by running "sqfscat" on the Squashfs
   image

2. Create a new "R" pseudo-definition, with the data contained
   within the pseudo-file itself.

Both options represent a trade-off between space-overhead and
I/O speed.

Generating a pseudo file using all "F .... sqfscat" commands
produces a very small file, but, it will cause Mksquashfs to
run slowly if most of the files are small (i.e. in the order
of 20 or so Kbytes).  This is because the overhead of forking
and executing a sub-process for that ammount of data repeatedly
will bottleneck Mksquashfs.

The new "R" pseudo-definition which stores the data inside
the pseudo file is fast, but, the pseudo file obviously
will become very large.

The inline_threshold and no_threshhold variables control
which technique is used.  Any file smaller than the
inline_threshold value is stored inside the pseudo file,
and any file larger is obtained via executing sqfscat.

If no_threshhold is set, then the files are always stored
inline, which is the case in this commit.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>