Unsquashfs: Remove use of "F" pseudo-def in -pseudo-file option
Standardise on the new inline data "D" pseudo definition to
store and generate regular files in the pseudo file.

This is because running sqfscat on the original Squashfs
filesystem is considered "dangerous", in that it makes
the pseudo-file dependent on the original filesystem
being available.

I think this "hidden" dependency on the original filesystem could
cause issues such as data loss, because many people will think the
pseudo-file is everything that is needed to recreate the
filesystem, and then they will discover they need the original
filesystem, but at which time it may be too late.

If something can be easily misunderstood or misused, it is
always better to not to do it.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>