Mksquashfs: Don't complain if sources and "/" pseudo def given
Sources always take precedence over pseudo files (a pseudo-file
will be ignored if it already exists in the sources).  So just
ignore any "/" pseudo-def if sources are present on the
command line.

This should reduce user confusion when using a pseudo file
output via the Unsquashfs -pseudo-file option, which in the
future *will* contain a definition for "/".

If that pseudo file is mixed with other sources, which is
a probable use-case, we don't want Mksquashfs to unexpectedly
exit because of a "/" definition conflict, and instead do
the more sensible action of ignoring it.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>