Mksquashfs: Implement -help (and -h) which outputs to stdout
An issue raised on GitHub (https://github.com/plougher/squashfs-tools/issues/111)
complains that when using the -help option, the help text is
output to stderr and not stdout.

It just so happens that Mksquashfs has never had an -help option!

Instead I implemented an output of the options text when
an error is found in the parsing of the command line,
for example an unexpected option, too few arguments
supplied etc.  Because this is an error situation, the
output of the options text is to stderr.

Once you have that functionality in place, you realise
you don't have to implement an explicit "-help" option.
Any -help option will be treated as an unrecognised
option, and it will produce the options help text - in
almost exactly the same way as if -help was implemented.

Why add more code and complexity, when the existing
functionality does what you want anyway.

The essential problem here is that the options text has grown
over the years, and so it is now more than a screen-full of
information.  Anyone attempting to pipe the output to a
utility such as "less" will discover the output is to stderr,
and not to stdout.

The solution to this problem is to finally implement
an "-help" option which explicity outputs to stdout,
rather than stderr.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>