Mksquashfs: Make entry->cache NULL on dummy file buffers
A  dummy file buffer is used to pass back the
result of a fragment duplicate being found by the
"front end fragment processing threads".  This
isn't a "real" filembuffer obtained from the cache,
and so it couldn't be freed by calling cache_block_put()
in the normal way.

Previously the code used a flag to indicate the buffer
should be simply freed, rather than calling cache_block_put().

Instead specially mark the dummy file buffer with a cache
field of NULL - which no "real" file buffer will have - and
check for that in cache_block_put().

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>