Mksquashfs: Delete inode if nlink count goes to zero
If an action causes a previously created file to be
deleted, the nlink count of the associated inode needs
to be decremented.

Previously the inode would be left orphaned if the nlink
count went to zero, which is harmless, but wastes memory.

Add a new routine which deletes the inode if the nlink
count goes to zero.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>