Mksquashfs: Move progress bar size update to inode creation
Previously the overall size of the progress bar (i.e. the
total number of blocks to be compressed) was computed
as part of the inode number allocation in dir_scan6(),
just before creating the filesystem in dir_scan7().

This means the current size of the progress bar is unknown
*while* scanning the source filesystem in dir_scan1(), because
it isn't computed until dir_scan6().

Due to the progress bar being zero size while scanning the
source filesystem it can't be displayed at this point.

For very large source filesystems (or ones that are slow to scan,
i.e. a network filesystem or on slow media) this means there can
be a while before Mksquashfs displays the progress bar.

So move computation of the progress bar size to inode
creation, while scanning the source filesystem in
dir_scan1().

Obviously, now the progress bar computation is done
*before* actions which can delete files, we need
to decrease the progress bar as part of dec_nlink_inode().

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>