tar: Don't report checksum error at end of tar file
A tar file should have two 512 zero-byte filled
headers at the end of the file.

Previously that wasn't dealt with, and such a header
would cause the tar file parser to report a checksum
error.

Deal with this in this commit.

Also add an ERROR message if we get an unexpected EOF
when reading the tar file header.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>