Mksquashfs: Enable progress bar in dir_scan1()
Move the enablement of the progress bar to dir_scan1(),
when the source directories are scanned.

As the source directory scanning is usually fast, on the
order of a couple of seconds, most people will notice
no difference in behaviour.

But where the source directories are extrememly large (or
extremely slow to read, being on a network drive or slow
physical media) this change means Mksquashfs will display
the progress bar while scanning the source directories,
rather than being completely silent.

In this case the progress bar will remain at 0% compressed,
but, the total amount of blocks will increase, giving
a visual indication that Mksquashfs is making progress
scanning the source directories.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>