xattrs: Check for tar files when reading xattrs
Tarfiles like pseudo files don't exist in the real
filesystem, and so the xattr code should treat them
as such, and not try to read xattrs from them.

As the "tar_file" pointer is in a union with the
"pseudo" pointer in the inode_info structure, this
effectively meant the existing IS_PSEUDO() check
should work for tarfiles as well - and mostly does.

Unfortunately, tar files may not define a directory in
a tarfile path, and only explicitly define the leaf
file.

In this case the tar code produces a default directory
definition, but one which doesn't have a tar_file structure
(because it was never read by the tar file parsing
code), and so tar_file is NULL.

This breaks the above xattr code.  Two ways of fixing
it are to produce a dummy tar_file, which seems wasteful
of memory, and the way which is done here to extend
the xattr check to check for tarfiles as well
(tarfiles need an additional flag to distinguish between
them and pseudo files where the different is important).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>