tar: Add value and test macro for hard links
Hard links obviously do not exist as a file type
in Unix/Linux.  This is because hard links are
resolved, and the directory entry holds an inode
reference to whatever the hard link resolves to.
If the hard link resolves to an S_IFREG type then
that is the type etc.

Unfortunately, tar files do have an hard link type,
and this has to be represented and stored, until it
can be resolved to an actual type.

It is pointless not to map tar types to Unix/Linux
types (struct stat) at the point of parsing
the tar file.  So we need someway of representing
Hard Links using that infrastructure, until it
can be resolved.

The solution here is to use the existing S_IFMT
definition as the hard link type, via a S_IFHRD
definition which is defined as S_IFMT.

Also add an S_ISHRD macro to test for a "hard link"
type.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>