Mksquashfs: write recovery file to $HOME
When appending to an existing filesystem, Mksquashfs
first writes a "recovery" file, so if Mksquashfs or the
computer crashes, it should be possible to recover
the original filesystem.

This recovery file has to be created and written
somewhere.  The obvious and usual place is in /tmp,
but this is not a good idea for a recovery file, as
many distributions clean /tmp when booting, which
will happen after a computer crash.

So traditionally Mksquashfs has always tried to create
the file in the current working directory, as this is
probably writeable and not /tmp.  The user can and
should always change to a writeable directory before
running Mksquashfs in append mode.

But the new -no-strip (and -tarstyle) option breaks
this behaviour.  These new options mean Mksquashfs
always includes the source  pathname components in
the output filesystem, and so the current working
directory directly affects what's stored.  If you
want "phillip/" to be stored in the root directory
your current working directory has to be /home, which
isn't writeable by a normal user.

In other words, the normal command line, which allows
Mksquashfs to be run from any writeable directory is:

% mksquashfs /home/phillip sqfs -keep-as-directory

that with the new -no-strip behaviour becomes

% cd /home; mksquashfs phillip sqfs -no-strip

The solution in this patch is to move to creating
the recovery file in the home directoy ($HOME).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>