Mksquashfs: add code to exclude destination file from input
Mksquashfs since the very early days (about 19 years),
has always deliberately created the destination file before
processing the exclude files.  Why you may ask?

This is to allow the user to exclude the destination file
in situations where the destination file will be picked
up as part of the input.  For example giving "example-dir"
as the source, and then creating the destination file
in that directory.

This exclusion was never done automatically, as users
were expected to "do the right thing", and Mksquashfs
concentrated on doing exactly what you told it to do.

But since those early days attitudes have changed.  It is
now generally assumed if the application allows you to
do something silly, it is a bug.  Additionally there
has arisen a whole class of exploits which try to
trigger undesirable behaviour by unexpected combinations.

So it well past time to always enforce explicit exclusion
of the destination file from the input.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>