Mksquashfs: fix failure to truncate when appending aborted
When appending to an existing Squashfs filesystem, if it is
aborted (by typing ^C twice), the original filesystem is
restored.

However, if the destination file had grown in length due to
appending, then the destination file would not be truncated
to the original length.

This patch fixes this by calling ftruncate() when restoring
the filesystem.

Note ftruncate() must be called, and then the file padded
with zeroes.  Ftruncate on its own will pad to the correct
length, but it won't be zero padded, and will simply have the
values already there.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>