Mksquashfs: behave like cpio, and read files from stdin
Whilst reviewing and testing the new -tarstyle source
handling code, I noticed it had added almost all the
necessary functionality to do "cpio style" reading
of the files from standard in (stdin).

Specifically, the differences between tar and cpio
handling are that the files to be added are (obviously)
read from standard in, and that directories are
not recursively scanned, and their contents added.  Every
file to be included has to explictly specified.

With tarstyle handling the code already has to merge the
input sources into a directory structure if they share
pathnames, and this code will also merge into the necessary
directory structure a cpio style list of files.

So the additional code needed is code to read filenames
from standard in, and a flag to not recursively
scan directories, as their content will be explictly
added.

This is probably a useful addition to Mksquashfs, as it
now allows various "find" based filtering of files
which is difficult to do otherwise, i.e.

% find . -name "*.c" | mksquashfs - img.sqfs -cpiostyle

will only add the *.c files.  This can be done with
file exclusion, but, it is often easier to achieve
what you want with find.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>