tar: handle Pax mtime keyword=value pair
I had hitherto not implemented Pax mtime decoding
because it was thought unnecessary, as the later
tar header invariably defines it at the lower
resolution, which is adequate for Mksquashfs.

But I have come across a "test tarfile" which
defines it in the extended Pax header, and then
encodes mtime as 0 in the later tar header.

This is obviously a "nasty" tarfile where the situation
is unlikely to be met in "real life", but, Pax mtime
parsing is easily added, and which eliminates the
failure to correctly decode that tarfile.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>