progressbar: don't print so many blank lines
The progressbar doesn't output a newline for obvious reasons
(this is to allow the progressbar to be updated).

What this means is whenever Mksquashfs wants to output
a message or error when the progressbar is enabled, it
has to output a newline first, to ensure it appears on
the line after the progressbar.

The problem here is if lots of messages and/or errors
are printed, they appear with a blank line between them,
because only the first line needed to output a newline
to ensure it appeared after the progressbar.  This looks
messy, wastes screen area, and increases the likelihood
the message or error will scroll off the screen before
it is noticed.

Add logic to only print a newline when necessary.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>