Mksquashfs: update directory size check
Issue https://github.com/plougher/squashfs-tools/issues/118
on Github complains the code is checking the directory
size against a maximum limit of 128 Mbytes (2^27 - 1).

This was the directory size limit with the old Squashfs 3.x
layout.  The code should have been updated to check for
the 4.0 layout limit of 4 Gbytes (2^32 - 1).

Do this now.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>