actions: process prune action for symlinks in a separate scan
Process prune actions for symlinks before processing prune
actions for other files.

This fixes the problem that the following intuitive action can
often fail.

prune@name(*.[ch])||readlink("name(*.[ch])")

This means delete any files named *.[ch] and delete any symlinks
that resolve to files named *.[ch].   This avoids leaving
any dangling symlinks.

The problem is often the file named *.[ch] can be deleted
*before* symlinks, leaving the symlink unresolvable, which
results in readlink returning FALSE (you can't test the
attributes of a file that can't be resolved).

This *can* be solved by testing whether the symlink is
resolvable, which deals with the fact the file may
have been deleted, or

prune@name(*.[ch])||!exists||readlink("name(*.[ch])")

But this will pick up symlinks that are dangling for
other reasons.

This commit solves the issue by running the prune
action on symlinks as a separate phase before
running the prune action on other files.  So you do not
get the situation a symlink has become unresolvable
because the file it points to has been pruned.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>