mksquashfs: document where to read the README, USAGE and ACTIONS-README
It is often clear that very few people read the latest release
README, or the USAGE guide.  So basically the only information
they read is the inbuilt help text (which is why I try to make it
as informative as possible within the constraints of a help text).

As far as I can tell the reason for this is because most people
get the Squashfs-tools from their distribution, and so they
don't see or even know there is more docmentation, because they
don't download the release tarball or go to the source code repository.

So I have decided to add URL links to the latest version of the
documentation (in the source code repository) to the inbuilt
help text.

Hopefully this will mean more people become aware there
is additional documentation available.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>