mksquashfs: handle options with arguments
When scanning the command line for options, if you hit
any options with arguments, you should skip the argument.
Otherwise there is a small possibility that an argument may
be mistaken for an option.  Obviously, most arguments don't
look like options, but, this could happen (especially
deliberately) where filenames are concerned.

The existing code did check for options with arguments, but,
this was ad-hoc, and has now become out of date with all the
new options added.

So replace this ad-hoc code with a new function.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>