progressbar: this can now shrink to zero
Previously the progress bar was not displayed whilst
the filesystem was being scanned.  This meant on very
large filesystems (or slow filesystems, the result is
the same), Mksquashfs would display nothing for a while,
whilst the filesystem was being scanned.

Now the progress bar is displayed whilst the filesystem
is being scanned.  But what this means is the progress bar
can later shrink if files are deleted due to pruning.

If the progress bar shrinks to zero (i.e. everything pruned),
it stops being displayed, leaving an out of date
progress bar.

Obviously pruning everything, leaving an empty filesystem
is pointless.  But, it is a valid thing for someone to
do if ultimately pointless, and Mksquashfs should behave
correctly in this case.

So fix the case that the progress bar is not displayed
when it has shrunk to zero size.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>