sqfstar: make images by default *not* exportable
Filesystem images being exportable by default is
a legacy setting in Mksquashfs, which has been there
since NFS exporting was implemented.

Back in about 2006, a new Squashfs use-case arose where
people wanted Squashfs to be exportable via NFS.  At
that time Squashfs wasn't exportable via NFS, and I duly
implemented it for the next release.

When it came to deciding whether to make it default or
not, I choose to make it default.  This is even though I
personally prefered to make it non-default as it
increased data overhead for the majority that didn't
need it.

The reason I chose to make it default was human nature.
I was strongly convinced that if I made it a non-default
option, I would still get people complaining Squashfs
didn't support NFS exporting, because they expected
it to be default behaviour (like every other filesystem
that supports NFS exporting).  In otherwords if it
didn't support NFS exporting by default, they wouldn't
look for an option to enable it, because that is
completely different to how other filesystems work.

In other other words if you make something non-default
that people expect to be default, don't be surprised
if they don't look for an option to enable it, and
instead continue to assume it doesn't support it.

As that is essentially an own-goal, I was more or less
forced to make it a default option.

But that time has gone, and NFS exporting is a use-case
which is AFAIK not very common now.  In any case
for the expected use-case of tar->squashfs it makes sense
not to generate an export table by default.

If the generation of an export table is wanted, you
can use the new -exports option.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>