mksquashfs: do not print exclude error in add_source()
If an exclude matches, do not print an error, as the exclude should be
expected.

When add_source() only handled the -tarstyle option whether or not
to print an error was a grey-area.  In the case of the -tarstyle option
add_source() is only used to process the leading components in the
source pathnames.  Arguably excluding an explicit leading component
is a error, because it doesn't make sense to explicitly include it
and then explictly exclude it.

But, non-anchored excludes could always have potentially matched here.

In any case, once add_source() was extended to handle -cpiostyle
file input, it means add_source() is always used to process the
input, and the input includes every file to be added
(cpiostyle does not recursively descend leaf directories).  In this
case it makes absoultely no sense to output an error if an exclude
matches.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>