sqfstar: only write to block device or existing file with -force
Make Sqfstar only write to a block device if -force is added.

This should be a double check to force people to check what they're
doing is correct, e.g. before overwriting an important filesystem.

Also allow existing files to be over-written if the -force
option is specified, rather than making people delete the file.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>