mksquashfs: fix logic which determines if sources in same directory
When using the -tarstyle and -cpiostyle[0] options, there is logic
which determines if all the sources are in the same directory, or not.

This is important, because if they are all in the same directory, that
directory's attributes (ownership, permissions, timestamp etc) can be
used to create the root directory.  Otherwise, if they are not in the
same directory, a root directory with default attributes has to be
used.

Adding support for leading "..", broke the logic.  This commit fixes
it, and simplifies it too.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>