Avoid use of INSTALL_DIR for symlink targets
In case INSTALL_DIR is overridden with a staged install location, using
INSTALL_DIR for the symlink target path prefix will yield an incorrect location
for the final installation.

Because the symlink itself is already installed to INSTALL_DIR, simply removing
the INSTALL_DIR prefix suffices as a fix.

Note that using $DESTDIR/$INSTALL_DIR where appropriate can avoid this type of
issue, but that can be considered a future enhancement.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>