Don't sanity check fragment table location if there are no fragments
If there are no fragments (as per the super block fragment count),
the fragment table is not accessed. In that case, makes no sense
to check the fragment table location or assume anything about it.
It could reasonably be set to e.g. SQUASHFS_INVALID_BLK like the
other fields which trips this check.

Furthermore, the kernel side implementation also does not perform
this check if the fragment count is zero. So with this check in
place, unsquashfs is no longer able to unpack images that the
kernel happily accepts as valid.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>