Unsquashfs: don't use max open file limit if larger than caches
Issue https://github.com/plougher/squashfs-tools/issues/125
complains that Unsquashfs aborts if the max open file limit is
larger than 134216713.

This is because Unsquashfs always uses the max open file limit
to base its queues on if it is defined, even if it is a
large value.

In fact Unsquashfs should only be be using the max open
file limit if it produces smaller queue sizes than
otherwise, as explained in the earlier comment

 * The maximum file open limit, however, affects the read-ahead
 * possible, in that for normal sizes of the fragment and data block
 * caches, where the incoming files have few data blocks or one fragment
 * only, the file open limit is likely to be reached before the
 * caches are full.  This means the worst case sizing of the combined
 * sizes of the caches is unlikely to ever be necessary.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>