Mksquashfs: add new "expert" category to options text
A pull request complains that there is no warning that
the -nopad option will create an unmountable filesystem.

Obviously, this is not entirely correct, as it won't
create an unmountable filesystem if the destination
is a block device.

Additionally, if a warning is to be added to -nopad,
then it should be added to every option that can
create an unmountable filesystem.  But, adding to
every option will soon get out of hand.

Also if you don't know what an option does, then you
shouldn't use it.

So, move all options that can create an unmountable
filesystem into a new "expert" category which flags
these as "expert" options which may create an
unmountable filesystem.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>