Unsquashfs: don't abort if there's unsupported xattrs in the filesystem
This fixes issue https://github.com/plougher/squashfs-tools/issues/120
where Unsquashfs 4.4 and Unsquashfs 4.5 will abort if there's
xattrs in the filesystem and Unsquashfs has been built without
xattr support.  Previous versions of Unsquashfs ignored the
xattrs in the filesystem in this case.

This is a regression (introduced in 2019) caused by extra sanity
checking (security hardening) of the filesystem.  Specifically,
to do the extra sanity checking, Unsquashfs needs to partially
parse the xattr table metadata, and without any xattr support,
the necessary code doesn't exist, and so Unsquashfs aborts
when calling the dummy no-xattr support function (which just
returns failure).

The solution is to partially read the xattr metadata in the
no-support function, to obtain the necessary information
required by the extra sanity checks.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>