sqfstar: initialise inode->tar_file to NULL if automatic directory
Issue https://github.com/plougher/squashfs-tools/issues/145
reports a Sqfstar crash under ASAN with an example tar file.

This issue is caused by an uninitialised value in the code
which deals with "automatically created directories".

A feature of tar files is a regular file may have directories
within the file pathname, which are not defined by the tar
file.

The example tar file is one of those.  It has a regular
file with the filename "000/ccc", where there is no
definition for the "000" directory within the tar file.

In this case Sqfstar automatically creates a "000" directory with
default values.  But, the directory doesn't have a tar_file
structure associated with it, because it wasn't read from
the tar file.  The code should have been setting tar_file
to NULL, but it didn't.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>