Sqfstar: Fix uninitialised value detected by MSAN build
Issue https://github.com/plougher/squashfs-tools/issues/146
reports a crash under MSAN caused by an uninitialised value.

This is due to an uninitialised fragment value in the file_buffer
structure passed to seq_queue_put().

In the tar implementation this functionality is not used, and so
simply initialise it to FALSE to get rid of the issue.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>