Unsquashfs: if "dest" dir already exists use it rather than aborting
For many people they either expect to create the destination
directory *before* calling Unsquashfs, or they have a pre-existing
directory (perhaps created by a build process).  In both
cases they expect Unsquashfs to extract into that directory.

But, Unsquashfs ever since the beginning has refused to extract
into pre-existing directories to avoid over-writing information.
This behaviour can be rather annoying if you've already got the
extract directory created for the above mentioned reasons.

There is a way to marry the dis-joint desire to extract into a
pre-existing directory, and to keep Unsquashfs' desire not
to alter and over-write pre-existing directories and files.

This is to make the top-level directory a special case, where
Unsquashfs does allow *that* directory to pre-exist.

In this case Unsquashfs will extract to a pre-existing directory (empty
or not), and the pre-existing checks will prevent over-writing of files
in that directory, or the descent into any directories within that
directory.

So it allows pre-existing directories to be extracted into, while
still preventing existing files being written to, or existing directories
being added to.

This seems to be better than the current situation.

Fixes issue https://github.com/plougher/squashfs-tools/issues/163

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>