Unsquashfs: Don't allow -dest <pathname> to be "" (empty)
If the <pathname> specified by the -dest option is "" (empty),
Unsquashfs aborts.

Trap this, and suggest they use ".", if they're trying to extract
into the current directory.

There is a temptation here to automatically replace "" with ".",
which would allow people to specify an empty pathname for the
current directory.

But, this is risky, as many scripts use variables, and an unexpectedly
undefined variable can resolve to "", i.e. "$example" where $example is
undefined.

In general if an error occurs, it is better to bail out with an error,
which will alert whoever needs to be alerted, rather than silently fixing it
which may be wrong and cause the error to be invisible.

Fixes https://github.com/plougher/squashfs-tools/issues/162

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>