Unsquashfs: allow destination to be empty when listing the filesystem
Since commit 91a2b9c it hasn't been possible to have an empty
destination.

However it has always been possible to specify an empty destination when
listing the contents of the filesystem. In fact some people or scripts
may do this to remove the default squashfs-root directory.

If that is the case, then the above commit will have broken the
behaviour.

So destination should be allowed to be empty when just listing the
filesystem.

Fixes https://github.com/plougher/squashfs-tools/issues/166

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>