Unsquashfs-3: Handle corrupted fs with directory loops
If a corrupted filesystem with directory loops is given to Unsquashfs,
it will endlessly recursively follow the loop, until either the pathname
exceeds the maximum path, and system calls fail (if extracting), or if
listing or walking pathname (sqfscat, follow-symlinks) until the
stack overflows.

To fix this requires tracking of directory inode lookup, and to abort if a
directory is looked-up more than once in a directory traversal.

The most time and space efficient way of doing this, is to implement
a bit-table, tracking directory inode numbers.

This commit adds the inode number lookup to Unsquashfs 3.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>