Unsquashfs: Update bit-table comment to mention hard-links
Update comment to state the bit-table not only traps directory
loops, but, also traps where a directory is invalidly hard-linked
to multiple locations in the directory tree.  A directory loop
obviously being a special case of multiple linking which is
self-referencial.

This is why a bit-table was used, rather than a stack, because a stack
will only detect directory loops.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>