Unsquashfs: Allocate inode lookup table dynamically
When walking the filesystem, we want to keep track of what inodes we've
already visited, as visting an inode more than once indicates a
hard-link, and when extracting, we want to create a hard-link to the
file already created.

To do this, obviously, requires a lookup table to be created *before*
walking the filesystem.

Originally the code used the inodes count in the Superblock, to create
the entire lookup table up-front before walking the filesystem.

Unfortunately, if the inodes count has become corrupted in the Superblock,
this can lead to Unsquashfs creating or trying to create a huge table (i.e.
16 Gigabytes for 2^31 inodes) and zero it out before doing anything else.

If the filesystem has other corruption, it means Unsquashfs hangs
creating this huge table, before aborting.

We'd like to verify the inodes count in the Superblock is correct *before*
allocating the table, but, the only way of doing that is to scan the entire
inode table up-front.  This itself will become a cause of slow start-up,
especially with Sqfscat and Unsquashfs where you're only extracting one or
a couple of files.

In fact up-front scanning of the inode table was *removed* in version
4.5 to make start-up faster in this situation.

The solution is implement a lookup table that dynamically grows, as the
filesystem is walked.  This means Unsquashfs never tries to allocate
a huge table up-front, and the lookup table only grows to the real
amount of inodes in the filesystem.

Fixes https://github.com/plougher/squashfs-tools/issues/159

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>