Unsquashfs: Fix conversion from array to function in lookup table
Commit 54cc891a88e3779433d850bb715805dab8a1ac25 moved the
inode lookup implementation from a directly coded flat
array to a two-level dynamic array via a function.

Inode numbers start from 1, and so to convert from inode number
to array index, 1 must be subtracted.

In the original directly coded flat array this was done
while indexing into the array.  In the new implementation
this is done within the functions.

So during the conversion, the substraction of 1 from the
inode number should have been removed, as it is already
handled in the functions.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>