Avoid passing giant values to read/write
Avoid casting from `long long` to `int` before calling a function which
takes `size_t`.

On a 64 bit system, In the case where
`bytes - count == (long long)INT_MAX + 1`, this will not be
`> SSIZE_MAX`, but when cast to `int`, it will become `INT_MIN`. When
`INT_MIN` is cast to `size_t`, it becomes an enormous number (much
greater than `SSIZE_MAX`, and much greater than `bytes - count`.

Perhaps worse, if `bytes - count == ((long long)INT_MAX + 1) * 2`, this
will become 0 when cast to `int`. `read` will then return `0`, which we
will treat as an early EOF.

In this case, because the overflow only occurs on values `> INT_MAX`,
and these functions don't seem to be ever called with values so large,
the change is largly academic, however, it has a side benefit of
silencing an error when compiled with clang about casting `SSIZE_MAX` to
`int` changing the value to `-1`