Limit read() size to Linux maximum of 0x7ffff000
Linux manpage READ(2) in Notes states the maximum transfer
size is 0x7ffff000.

Previously SSIZE_MAX was used in the code, but, the value of that is
unreliable on different compilers.  This produces code which does not
limit the read size to 0x7ffff000.

Due to the unreliability of SSIZE_MAX, this commit switches to
a hard coded value.

Fixes https://github.com/plougher/squashfs-tools/issues/171

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>