manpage scripts: Fix generation error when using dash
The scripts use /bin/sh to execute the scripts.  However
this may point to a number of different shell implementations,
depending on the distribution.

I have noticed there is a script incompatibility depending
on which shell is used for /bin/sh.

The Bash echo builtin recognises the "-e" option to process
backslashed characters (in this case "\n").

But the Dash (used by Ubuntu) echo builtin does not recognise the
"-e" option leading it to be included in the output, producing
an erroneous "-e" in the manual pages.

This could be worked around by making the scripts use Bash as
the shell, but, this will add a dependency on a shell which
may not be installed on some distributions, and it seems
churlish to require Bash to be installed, when it can be
worked around another way.

This commit removes the use of "echo -e" to enable an embedded
"\n" to be recognised, and replaces it with an explicit echo to
add a newline.

This more basic use of echo is compatible with the Bash and Dash
builtins, and should be compatible with other shells which
may be used.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>